bot modules moved to bot folder
