+++ bot/__init__.py
@@ -1,0 +1,1 @@
+from bot.jack import JackBot  # noqa F401
