--- core/__init__.py
@@ -1,1 +1,0 @@
-from core.telegram import BotTelegramCore  # noqa F401
